\name{alpha.not.beta-package}
\alias{alpha.not.beta-package}
\alias{alpha.not.beta}
\docType{package}
\title{
\packageTitle{alpha.not.beta}
}
\description{
\packageDescription{alpha.not.beta}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{alpha.not.beta}
\packageIndices{alpha.not.beta}
You want a letta, you got a letta 
}
\author{
\packageAuthor{alpha.not.beta}
}


\keyword{ package }
\seealso{
}
\examples{
alphabet()
}
