\name{alphabet}
\alias{alphabet}
\alias{alpha.not.beta}

\title{
 Gives a random letter
}
\description{
 Generates a random letter, how lucky
}
\usage{
alphabet()
}

\references{
big shout out to the greg, package wouldn't be possible without him
}
\author{
Akhil Ghosh
}

\examples{

a <- alphabet()

## The function is currently defined as
function ()
{
    return(sample(letters, size = 1))
  }
}

